add clipboard api and refactoring fullscreen, pointer-lock apis
