+++ packages/client/src/web-api/clipboard-api.ts
@@ -1,0 +1,9 @@
+class ClipboardAPI {
+  copy(value: string) {
+    navigator.clipboard.writeText(value);
+  }
+}
+
+const clipboardAPI = new ClipboardAPI();
+
+export default clipboardAPI;
