+++ packages/client/src/web-api/pointer-lock-api.ts
@@ -1,0 +1,9 @@
+class PointerLockAPI {
+  lockByElement(element: HTMLElement) {
+    element.requestPointerLock();
+  }
+}
+
+const pointerLockAPI = new PointerLockAPI();
+
+export default pointerLockAPI;
